requirements.txt
aiohttp
python-decouple
sqlalchemy
requests
sqlalchemy-utils
psycopg2-binary
asyncpg
asyncpgsa
